Move all createGroup and Group.can calls into one file
