--- packages/lesswrong/lib/collections/advisorRequests/permissions.ts
@@ -1,8 +1,0 @@
-import { membersGroup } from '../../vulcan-users/permissions';
-
-const membersActions = [
-  'advisorrequests.view.own',
-  'advisorrequests.new',
-  'advisorrequests.edit.own',
-];
-membersGroup.can(membersActions);
