+++ packages/lesswrong/lib/collections/bans/collection.ts
@@ -33,4 +33,9 @@ export const Bans: BansCollection = createCollection({
 
 addUniversalFields({collection: Bans})
 
+Bans.checkAccess = async (user: DbUser|null, document: DbBan, context: ResolverContext|null): Promise<boolean> => {
+  if (!user || !document) return false;
+  return userCanDo(user, 'bans.view')
+};
+
 export default Bans
