+++ packages/lesswrong/lib/collections/chapters/collection.ts
@@ -51,4 +51,11 @@ makeEditable({
 })
 addUniversalFields({collection: Chapters})
 
+Chapters.checkAccess = async (user: DbUser|null, document: DbChapter, context: ResolverContext|null): Promise<boolean> => {
+  if (!document) return false;
+  // Since chapters have no userIds there is no obvious way to check for permissions.
+  // We might want to check the parent sequence, but that seems too costly, so for now just be permissinve
+  return true
+};
+
 export default Chapters;
