+++ packages/lesswrong/lib/collections/conversations/collection.ts
@@ -46,4 +46,9 @@ addUniversalFields({
   createdAtOptions: {canRead: ['members']},
 })
 
+Conversations.checkAccess = async (user: DbUser|null, document: DbConversation, context: ResolverContext|null): Promise<boolean> => {
+  if (!user || !document) return false;
+  return document.participantIds?.includes(user._id) ? userCanDo(user, 'conversations.view.own') : userCanDo(user, `conversations.view.all`)
+};
+
 export default Conversations;
