+++ packages/lesswrong/lib/collections/lwevents/collection.ts
@@ -36,4 +36,10 @@ addUniversalFields({
   createdAtOptions: {canRead: ['members']},
 });
 
+LWEvents.checkAccess = async (user: DbUser|null, document: DbLWEvent, context: ResolverContext|null): Promise<boolean> => {
+  if (!user || !document) return false;
+  if (document.name === "gatherTownUsersCheck") return true
+  return userOwns(user, document) ? userCanDo(user, 'events.view.own') : userCanDo(user, `events.view.all`)
+};
+
 export default LWEvents;
