+++ packages/lesswrong/lib/collections/multiDocuments/collection.ts
@@ -1,12 +1,11 @@
 import { makeEditable } from "@/lib/editor/make_editable";
 import schema from "./schema";
 import { ensureIndex } from "@/lib/collectionIndexUtils";
-import { membersGroup, userIsAdmin, userOwns } from "@/lib/vulcan-users/permissions";
+import { userIsAdmin, userOwns } from "@/lib/vulcan-users/permissions";
 import { canMutateParentDocument, getRootDocument } from "./helpers";
 import { makeVoteable } from "@/lib/make_voteable";
 import { addSlugFields } from "@/lib/utils/schemaUtils";
 import { createCollection } from "@/lib/vulcan-lib/collections.ts";
-import { getCollection } from "@/lib/vulcan-lib/getCollection.ts";
 import { addUniversalFields } from "@/lib/collectionUtils";
 import { getDefaultMutations } from "@/lib/vulcan-core/default_mutations.ts";
 import { getDefaultResolvers } from "@/lib/vulcan-core/default_resolvers.ts";
@@ -91,13 +90,6 @@ MultiDocuments.checkAccess = async (user: DbUser | null, multiDocument: DbMultiD
 
   return true;
 };
-
-membersGroup.can([
-  'multidocuments.smallDownvote',
-  'multidocuments.bigDownvote',
-  'multidocuments.smallUpvote',
-  'multidocuments.bigUpvote',
-]);
 
 makeVoteable(MultiDocuments, {
   timeDecayScoresCronjob: false,
