+++ packages/lesswrong/lib/collections/notifications/collection.ts
@@ -36,4 +36,9 @@ addUniversalFields({
   createdAtOptions: {canRead: [userOwns]},
 })
 
+Notifications.checkAccess = async (user: DbUser|null, document: DbNotification, context: ResolverContext|null): Promise<boolean> => {
+  if (!user || !document) return false;
+  return userOwns(user, document) ? userCanDo(user, 'notifications.view.own') : userCanDo(user, `conversations.view.all`)
+};
+
 export default Notifications;
