+++ packages/lesswrong/lib/collections/reports/collection.ts
@@ -1,6 +1,5 @@
 import schema from './schema';
-import { userCanDo, membersGroup } from '../../vulcan-users/permissions';
-import { sunshineRegimentGroup } from '../../permissions';
+import { userCanDo } from '../../vulcan-users/permissions';
 import { createCollection } from '../../vulcan-lib/collections';
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
@@ -22,20 +21,6 @@ addUniversalFields({
     canUpdate: ['admins'],
   },
 });
-
-const membersActions = [
-  'reports.new',
-  'reports.view.own',
-];
-membersGroup.can(membersActions);
-
-const sunshineRegimentActions = [
-  'reports.new',
-  'reports.edit.all',
-  'reports.remove.all',
-  'reports.view.all',
-];
-sunshineRegimentGroup.can(sunshineRegimentActions);
 
 Reports.checkAccess = async (user: DbUser|null, document: DbReport, context: ResolverContext|null): Promise<boolean> => {
   if (!user || !document) return false;
