+++ packages/lesswrong/lib/collections/reviewVotes/collection.ts
@@ -1,5 +1,5 @@
 import schema from './schema';
-import { membersGroup, userCanDo } from '../../vulcan-users/permissions';
+import { userCanDo } from '../../vulcan-users/permissions';
 import { createCollection } from '../../vulcan-lib/collections';
 import { sunshineRegimentGroup } from '../../permissions';
 import { addUniversalFields } from "../../collectionUtils";
@@ -13,19 +13,6 @@ export const ReviewVotes: ReviewVotesCollection = createCollection({
 });
 
 addUniversalFields({collection: ReviewVotes})
-
-const membersActions = [
-  'reviewVotes.new',
-  'reviewVotes.view.own',
-];
-membersGroup.can(membersActions);
-
-const sunshineRegimentActions = [
-  'reviewVotes.edit.all',
-  'reviewVotes.remove.all',
-  'reviewVotes.view.all',
-];
-sunshineRegimentGroup.can(sunshineRegimentActions);
 
 ReviewVotes.checkAccess = async (user: DbUser|null, document: DbReviewVote, context: ResolverContext|null): Promise<boolean> => {
   if (!user || !document) return false;
