--- packages/lesswrong/lib/collections/subscriptions/permissions.ts
@@ -1,7 +1,0 @@
-import { membersGroup } from '../../vulcan-users/permissions';
-
-const membersActions = [
-  "subscriptions.new"
-];
-
-membersGroup.can(membersActions);
