--- packages/lesswrong/lib/collections/tagRels/permissions.ts
@@ -1,10 +1,0 @@
-import { membersGroup } from '../../vulcan-users/permissions';
-
-const votingActions = [
-  'tagrels.smallDownvote',
-  'tagrels.bigDownvote',
-  'tagrels.smallUpvote',
-  'tagrels.bigUpvote',
-]
-
-membersGroup.can(votingActions);
