+++ packages/lesswrong/lib/collections/tags/permissions.ts
@@ -1,4 +1,3 @@
-import { membersGroup } from "@/lib/vulcan-users/permissions.ts";
 import { canVoteOnTag } from "../../voting/tagRelVoteRules";
 import { CoauthoredPost } from "../posts/helpers";
 
@@ -9,10 +8,3 @@ export const shouldHideTagForVoting = (
   tag: TagWithVotePermissons,
   post: {userId?: string} & CoauthoredPost|null,
 ) => canVoteOnTag(tag?.canVoteOnRels, user, post, 'smallUpvote').fail;
-
-membersGroup.can([
-  'tags.smallDownvote',
-  'tags.bigDownvote',
-  'tags.smallUpvote',
-  'tags.bigUpvote',
-]);
