--- packages/lesswrong/lib/collections/userEAGDetails/permissions.ts
@@ -1,8 +1,0 @@
-import { membersGroup } from '../../vulcan-users/permissions';
-
-const membersActions = [
-  'usereagdetails.view.own',
-  'usereagdetails.new',
-  'usereagdetails.edit.own',
-];
-membersGroup.can(membersActions);
