--- packages/lesswrong/lib/collections/userJobAds/permissions.ts
@@ -1,8 +1,0 @@
-import { membersGroup } from '../../vulcan-users/permissions';
-
-const membersActions = [
-  'userjobads.view.own',
-  'userjobads.new',
-  'userjobads.edit.own',
-];
-membersGroup.can(membersActions);
