--- packages/lesswrong/lib/collections/userRateLimits/permissions.ts
@@ -1,10 +1,0 @@
-import { adminsGroup } from "../../vulcan-users/permissions";
-import { sunshineRegimentGroup } from "../../permissions";
-
-const actions = [
-  "userratelimits.new",
-  "userratelimits.edit.all",
-];
-
-adminsGroup.can(actions);
-sunshineRegimentGroup.can(actions);
