+++ packages/lesswrong/lib/vulcan-users/permissions.ts
@@ -17,13 +17,13 @@ class Group {
     this.actions = this.actions.concat(actions);
   }
 
-  cannot(actions: string|string[]) {
+  _cannot(actions: string|string[]) {
     actions = Array.isArray(actions) ? actions : [actions];
     this.actions = _.difference(this.actions, actions);
   }
 }
 
-export const userGroups: Record<string,Group> = {};
+const userGroups: Record<string,Group> = {};
 
 
 // Create a new group
@@ -31,6 +31,10 @@ export const createGroup = (groupName: string): Group => {
   userGroups[groupName] = new Group();
   return userGroups[groupName];
 };
+
+export function getAllUserGroups() {
+  return userGroups;
+}
 
 export type PermissionableUser = UsersMinimumInfo & Pick<DbUser,
   "groups" |
@@ -362,36 +366,3 @@ export const userCanUpdateField = <N extends CollectionNameString>(
   }
   return false;
 };
-
-////////////////////
-// Initialize     //
-////////////////////
-
-// initialize the 3 out-of-the-box groups
-export const guestsGroup = createGroup('guests'); // non-logged-in users
-export const membersGroup = createGroup('members'); // regular users
-
-const membersActions = [
-  'user.create',
-  'user.update.own',
-  // OpenCRUD backwards compatibility
-  'users.new',
-  'users.edit.own',
-  'users.remove.own',
-];
-userGroups.members.can(membersActions);
-
-export const adminsGroup = createGroup('admins'); // admin users
-
-const adminActions = [
-  'user.create',
-  'user.update.all',
-  'user.delete.all',
-  'setting.update',
-  // OpenCRUD backwards compatibility
-  'users.new',
-  'users.edit.all',
-  'users.remove.all',
-  'settings.edit',
-];
-userGroups.admins.can(adminActions);
